requirements.txt
fastapi
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
google-genai>=0.8.0
nest_asyncio
streamlit
bcrypt
bcrypt==4.1.2
streamlit==1.39.0
pandas==2.2.3
python-dotenv==1.0.1
passlib==1.7.4
google-generativeai==0.8.3
streamlit-lottie==0.0.5
requests==2.32.3
nest_asyncio==1.6.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
tweepy==4.14.0
requests==2.32.3
passlib[bcrypt]==1.7.4
cryptography==43.0.1
pydantic==2.9.2





